Limit udp_proxy source files to one logger ID
